added java for vim and c++ for vim

+++ install_mac.sh
@@ -2,7 +2,8 @@
 
 
 ## ANDROID
-if [ ! -e ./android-sdk ]; then
+if [ ! -e ~/opt/android-sdk ]; then
+	cd ~/opt
     wget http://dl.google.com/android/android-sdk_r23.0.2-macosx.zip
     unzip android-sdk_r23.0.2-macosx.zip
     rm android-sdk_r23.0.2-macosx.zip
@@ -13,6 +14,13 @@ if [ ! -e ./android-sdk ]; then
     ln -s android-ndk-r10b android-ndk
 fi
 
+## Eclipse
+if [ ! -e ~/opt/eclipse ]; then
+	cd ~/opt
+	wget http://mirror.csclub.uwaterloo.ca/eclipse/technology/epp/downloads/release/luna/SR1/eclipse-java-luna-SR1-macosx-cocoa-x86_64.tar.gz
+    tar -zxf eclipse-java-luna-SR1-macosx-cocoa-x86_64.tar.gz
+fi
+
 # HomebrewA
 brewbin=`which brew`
 if [ ! -e $brewbin ]; then
@@ -20,9 +28,16 @@ if [ ! -e $brewbin ]; then
 fi
 
 ## fish-shell
-brew install fish
+if [ ! -e `which fish` ]; then
+    brew install fish
+fi
 #chsh -s /usr/local/bin/fish
 
 ## VIM OSX
 #brew install vim --override-system-vim --without-python
+if [ ! -e `which nvim` ]; then
+    brew tap neovim/homebrew-neovim
+    brew install --HEAD neovim
+fi 
+
